Compute workspace statistics in relocation
