+++ front/lib/api/poke/plugins/workspaces/compute_statistics.ts
@@ -1,20 +1,6 @@
-import assert from "assert";
-
-import { computeDataSourceStatistics } from "@app/lib/api/data_sources";
 import { createPlugin } from "@app/lib/api/poke/types";
-import { DataSourceResource } from "@app/lib/resources/data_source_resource";
-import { concurrentExecutor } from "@app/lib/utils/async_utils";
-import { Err, maxFileSizeToHumanReadable, Ok, removeNulls } from "@app/types";
-
-type WorkspaceStats = {
-  dataSources: {
-    name: string;
-    text_size: string;
-    document_count: number;
-  }[];
-  document_count: number;
-  text_size: number;
-};
+import { computeWorkspaceStatistics } from "@app/lib/api/workspace_statistiques";
+import { Err, Ok } from "@app/types";
 
 export const computeWorkspaceStatsPlugin = createPlugin({
   manifest: {
@@ -29,57 +15,16 @@ export const computeWorkspaceStatsPlugin = createPlugin({
       return new Err(new Error("Workspace not found."));
     }
 
-    // Exclude conversation data sources on purpose. It would be too heavy to compute.
-    const dataSources = await DataSourceResource.listByWorkspace(auth, {
-      includeDeleted: true,
-    });
-
-    const results = await concurrentExecutor(
-      dataSources,
-      async (dataSource) => computeDataSourceStatistics(dataSource),
-      { concurrency: 10 }
-    );
-
-    const hasError = results.some((r) => r.isErr());
-    if (hasError) {
-      return new Err(
-        new Error("Error computing statistics.", {
-          cause: removeNulls(
-            results.map((r) => (r.isErr() ? r.error.message : null))
-          ),
-        })
-      );
+    const statsRes = await computeWorkspaceStatistics(auth);
+    if (statsRes.isErr()) {
+      return new Err(statsRes.error);
     }
 
-    const stats = results.reduce<WorkspaceStats>(
-      (acc, r) => {
-        // Errors are filtered out above.
-        assert(r.isOk());
-
-        const { name, text_size, document_count } = r.value.data_source;
-
-        return {
-          text_size: acc.text_size + text_size,
-          document_count: acc.document_count + document_count,
-          dataSources: [
-            ...acc.dataSources,
-            {
-              name,
-              text_size: maxFileSizeToHumanReadable(text_size, 2),
-              document_count,
-            },
-          ],
-        };
-      },
-      { text_size: 0, document_count: 0, dataSources: [] }
-    );
+    const stats = statsRes.value;
 
     return new Ok({
       display: "json",
-      value: {
-        ...stats,
-        text_size: maxFileSizeToHumanReadable(stats.text_size, 2),
-      },
+      value: stats,
     });
   },
 });
